requirements.txt
# requirements.txt

streamlit
Pillow
cloudinary
python-dotenv
openai>=1.0.0
pymongo